logo click redirects to landing, organized component into folder
